Lab2/requirements.txt
numpy==1.24.3
Flask==2.3.3
gunicorn==20.1.0
python-dotenv==1.0.0
tensorflow==2.13.0
keras==2.13.0
scikit-image==0.21.0
opencv-python==4.8.0.74
Pillow==10.0.0
matplotlib==3.7.2
datasets==2.14.0
scipy==1.11.2
pytest==7.4.0
pytest-flask==1.2.0
pytest-cov==4.1.0
requests==2.31.0
tqdm==4.66.1
scikit-learn==1.3.1